Renamed KeyFile to VaultStore and added more documentation
